recipe_app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.0,<4.0
psycopg2-binary